lambda_create_ig_stories/requirements.txt
boto3==1.34.153
pillow==10.4.0